Add Card and Navbar components with corresponding styles and stories
This commit includes the addition of Card and Navbar components along with their corresponding SCSS files. Necessary changes were made in the "components/index.scss" and "tokens/index.scss" to import the new components. Storybook stories were created for the new components and changes in the package.json file reflect a version increment.