Add modal component and update CSS token values
A modal component has been added to the CSS library. There's been an adjustment to the color tokens used in this update, and semantic tokens for screen width and background overlay were introduced. Furthermore, some unused tokens were deleted to improve the organization of the code. Lastly, inputs and modal components were updated according to the new token values.