Create Section Header component (#4)
* Create Section Header component

The inner `__content` div ensures the length of the divider is set
relative to the length of the text. The root div ensures the component
is still a block.

* Refactor section header component styles and remove unused import

The code changes involve transforming the 'section-header' component styles to be more maintainable and explicit. This includes using Sass '@use' rule instead of '@import', updating the gap for 'section-header' content and using variables for the styles of its divider. The component's unused import in the storybook has also been removed. Additionally, typography assertions have been made more precise.

---------

Co-authored-by: Demerzel <demerzel.olivaw@gmail.com>