Release/0.0.8 alpha (#7)
* Include variables when importing main SASS module (#3)

* Use `@use` over `@import` to import SASS modules

* Use `@forward` to allow importing vars via main

Only importing this library's main module doesn't give you access to the
variables defined within unless tokens and components are imported in
the main module with `@forward`.

* Fix leftover incorrect import path

* Update imports in README

* Add section to README regarding namespacing

* Redesigned Button component. Added Choicebox and Radio tokens. (#6)

* Create Section Header component (#4)

* Create Section Header component

The inner `__content` div ensures the length of the divider is set
relative to the length of the text. The root div ensures the component
is still a block.

* Refactor section header component styles and remove unused import

The code changes involve transforming the 'section-header' component styles to be more maintainable and explicit. This includes using Sass '@use' rule instead of '@import', updating the gap for 'section-header' content and using variables for the styles of its divider. The component's unused import in the storybook has also been removed. Additionally, typography assertions have been made more precise.

---------

Co-authored-by: Demerzel <demerzel.olivaw@gmail.com>

* v0.0.8-alpha release

---------

Co-authored-by: 11k <11k@users.noreply.github.com>