Fix minor issues with Card component (#15)
* Fix spacing when card contains empty sections

If a Card section is empty, the gap between it and the sections around
it will still exist, resulting in inconsistent spacing. Hiding empty
sections with `display: none` fixes the issue. Only extra and
description sections are relevant for this issue because they're the
only ones that can be empty in practice.

* Allow wrapping of contents in Card sections

A Card section that contains more badges or buttons than don't fit in a
single row results in overflow. Enabling wrapping fixes the issue and
makes responsiveness easier. Decreasing the row gap within sections
better indicates that the items within belong to the same section, even
if they wrap.

* Add changeset