Port drawer behavior from Dgg
With the old styles, setting margin-top on socials and action footers
was necessary in order to anchor the footers to the bottom of the
drawer. In Dgg, however, the layout is modified by wrapping all the
categories in a category group, and stretching the element to fill all
available space. This has the same effect and allows for scrolling of
the drawer items only, too.